+++ docpie/example/odd_even_example.py
@@ -1,0 +1,14 @@
+'''
+Usage: odd_even_example.py [-h | --help] (ODD EVEN)...
+
+Example, try:
+  odd_even_example.py 1 2 3 4
+
+Options:
+  -h, --help
+'''
+
+from docpie import docpie
+
+if __name__ == '__main__':
+    print(docpie(__doc__))